Feature: init startup time
